Persist work titles in cached index file
Titles have a habit of changing over time on MangaDex. Since this
program uses the title as the identifier for the work, a title
change would lead to re-downloading everything and leaving an old
duplicate folder behind. This patch adds a cached index file that
keeps track of the title associated with an ID locally. So even if
it changes on the site the local list will persist.