Fix chapters without groups not being downloaded
The check to filter out official scanlation groups inadvertently
ended up filtering out chapters that had no scanlation group
associations. This patch changes the query to a negated test that
filters only chapters that have an official group linked.

As a side benefit, the refactor removed the need to `uniq` the
output. This is thanks to using the `any()` filter.