Change chapter folders to include username instead of "no group"
This is an attempt to avoid collisions in the future. If multiple
users upload unattributed chapters and they happen to have the same
version number then only the first one would be saved. This patch
changes the folder naming convention to use the username of the
uploader if there are no associated scanlation groups.