Update existing chapter folder names if new version is identical
It is common for a chapter version number to get bumped without any
changes to the images. This was causing the program to create
redundant versioned folders with duplicate content.

This patch makes the script update existing folder names in that
scenario. It will check to see if there is a previous version of
the chapter already saved. Then if found it will diff the contents
with the latest version. When the content is identical it will
rename the existing folder to the latest version number.