Relocate title cache file to `$XDG_DATA_HOME` folder
This patch changes the title cache file location from the XDG cache
folder to the XDG data folder. It also includes a check to move the
file to the new location if one was found in the old location.

After reconsidering the file location, I think that the data folder
makes much more sense for this file. The title cache data cannot be
safely regenerated and is intended to be persistent. It is also
something that a user could want to backup and/or share between
computers. The cache folder is meant for things that are safe to
delete and can easily be regenerated. The XDG data folder seems
like a much better fit for this kind data.

A sound argument could be made to put it in the state folder, or
even the config folder. But, while it is user editable it's not
really user configuration, so the config folder feels like an
awkward fit. The state folder sort of makes sense, but to me this
subjectively feels more like "data" than "state".