Refactor chapter pages download to use single curl command
This patch replaces the manual page download looping with code to
generate a curl command that will download all of the page files
with one invocation.

This is a cleaner implementation, which lets curl handle more of
the mechanics of the download operation - such as rate limiting,
error handling, and retry logic. It also yields a noticeable speed
up improvement on my computer.

The curl program has an exponential backoff algorithm applied as
part of its retry logic. So this patch also removes the need for
the bc dependency. Which was used to perform a (very simplistic)
form of retry backoff.