Add handling for chapters that do not include chapter number
I found that it is possible for chapters in a series to be uploaded
without specifying a chapter number. It seems that this is intended
for "oneshot" manga which only have a single chapter. For a series
with multiple chapters this can really only happen if someone makes
a mistake when setting the metadata.

This patch adds code to check for chapters that do not specify a
chapter number, and skip them. For this script I cannot think of a
more intelligent way to handle this kind of mistake. Because I
do not want to pollute the download directory with some number of
unsortable "unknown chapter" folders.

Note that I had to modify the jq filter to include a string literal
"null" return using the alternative operator. If the first column
has an empty value then it is ignored as leading whitespace in any
code trying to handle the array, and will cause off-by-one errors.